Changed Storage class to allow multiple Storage objects to use the same file

+++ storage.py
@@ -3,10 +3,10 @@ from config import STORAGE_DIR
 
 class Storage:
   def __init__(self, filename: str):
-    self.file = Path(STORAGE_DIR) / filename
-    self.file.write_text('')  # Clear file and create if necessary
     storage_dir = Path(STORAGE_DIR)
-    if not storage_dir.exists(): storage_dir.mkdir(exist_ok=True)
+    if not storage_dir.exists(): storage_dir.mkdir(parents=True, exist_ok=True)
+    self.file = storage_dir / filename
+    if not self.file.exists(): self.file.write_text('')   # Create file if needed
 
 def store(self, data: str):
   return self.file.write_text(data)
